[chrome] Create empty target for arm64
workstation_*.arm64 builds are broken because there is no chrome browser
on arm64. Since we add chrome to the build in the product config we
don't have access to target_cpu to gate inclusion. To work around this
create an empty target for arm64.

Fixed: 108361

Change-Id: If3b498b051fc720783a9380c645d12e4df84241e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719212
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>