[storage][conformance] Enforce max read size of MAX_TRANSFER_SIZE.
Enforces the documented limit in fuchsia.io of MAX_TRANSFER_SIZE for any
transfer buffer. This is enforced on write by the fidl client code, but
for read the server should reject requests for larger sizes.

This found a bug in the cppvfs server where it would try to send an
oversized reply and hang the fidl connection.

Bug: 82674
Change-Id: I5694c5d2c953e055036efc34740aa831c233c40a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718999
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>