[hypervisor][x86] Fix invvpid after vmxoff in DirectGuest
The address space associated with a direct mode guest may be destroyed
after the final guest has been destroyed, and therefore VMX has been
turned off. This will then cause an exception to be raised.

To avoid this, we reset the VPID associated with the address space on
destruction. This is safe as on subsequent creation of a guest, we will
always issue an invept, and then on create of a VCPU we will issue an
invvpid.

Bug: 95763
Change-Id: I0e3239f3a6cd7953aa475b6dffab6f3b6ce5b728
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719215
Reviewed-by: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>