[f2fs] Count the i_block in the same way as a generic filesystems.
For compatibility, change the in-memory i_block to 512-byte sectors, the same as a generic filesystems.
In addition, the inode block is excluded from the in-memory i_block.
On the other hand, we calculate the block count in the on-disk i_block and keep the inode block included.

Change-Id: I6a3b80c4d2880e65ff5010a38cbb02eabadd4801
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708082
Fuchsia-Auto-Submit: Aaron Drew <ripper@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Drew <ripper@google.com>