[fuchsia-map] Create component and ffx plugin
The component called process_explorer gets all the processes on a running Fuchsia device, and for each of them it retrieves information about all the handles in the process. The data is then packaged as a string in JSON format and sent over a socket.

Component tested by using the Realm Builder library and mocking fuchsia.kernel.RootJob.

The ffx plugin, which is also called process_explorer, starts the component, receives the data about all processes as a string and prints it.

Change-Id: I1789f6119500853b12ee6f8dbaf3f84718806132
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699522
Reviewed-by: Benjamin Lerman <qsr@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Catalina Agachi <acatalina@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>