[bt][fastpair] Support pairing over LE & BR/EDR
Context: The remote device will initiate Fast Pair over LE. When the
peer is first discovered by the system, it is assigned a unique PeerId.
The remote device will then try to pair with us. In some cases, this is
over BR/EDR and as such, the system will assign a new unique PeerId to
the same remote device. Because the ids differ for the same peer,
`bt-fastpair-provider` will reject the pairing request.

Update the PairingManager to try and match the PeerId to an active Fast
Pair pairing procedure. This is a best-effort operation and likely to
succeed. It is unlikely that there are multiple ongoing pairing
procedures.

File fxbug.dev/107780 to see if we can simplify the implementation such
that only one Fast Pair pairing procedure is supported at a time.

Bug: 107780
Test: fx test bt-fastpair-provider-tests

Change-Id: I69a1cab05f682b421f645331441a3631a8307af8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716230
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>