[tools][fuzz] Make SSHInstanceCmd handle pipes more like exec.Cmd
Some differences in the way stdout/stderr pipes are treated by
ssh.Session vs exec.Cmd meant that we needed to Wait() on the remote
process to exit *before* indefinite reads on the process output could
receive EOFs and terminate. This is explicitly forbidden by exec.Cmd,
which requires reads to complete before calling Wait(), so we wrap the
SSH commands to behave more similarly.

This is in preparation for a following CL, which will use the same
Fuzzer management code for both local and remote processes.

Change-Id: Id557fe4d5c65e2e7aec9fe61aa7a9f06c7a83360
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717755
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>