[fidl][dynsuite] unknown interactions server tests.
Implement unknown interactions tests for the server in the DynSuite.
These tests are enabled for LLCPP, Natural Types, and Rust. For Go, they
are disabled because we have not implemented unknown interactions
support. For HLCPP they are disabled because unknown interactions
support is pending but not yet committed.

Bug: 88366

Change-Id: Ib947f56eb6fb23688512ce22e5573f7fb62ee6c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709902
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>