[ffx] Fix ephemeral target expiry
This CL fixes a bug in the FFX daemon's TargetCollection when handling
the expiration of emulation targets. In the original design, the target
address was stored in the ffx_config, so a daemon restart would retain
the entry. This entry would then be removed on daemon restart if the
entry had timed out. This led to an issue where an emulator that was
still running could be removed from the collection if the daemon
restarted, so the ffx config update was removed. At that point, an
ephemeral target would be removed from the daemon's in-memory
collection, but never ffx_config unless the user explicitly called `ffx
emu stop`.

This CL adds code to target expiry that will remove ephemeral targets
from ffx config as well as the in-memory collection. This code is
triggered only when the target's timeout has elapsed AND the target has
transitioned to a Disconnected state, indicating the emulator is no
longer (or never was) running. This is done by instantiating an ffx
config-backed ManualTargets data structure in the expiry loop (mirroring
the one in the TargetCollection protocol). Note that each ManualTargets
struct performs a read/update/write sequence for each operation, without
caching, so no consistency protocol is required.

Bug: 108351
Bug: 105550
Change-Id: I67cdc74585eddc4df9b13b38a7868f6b2fda8f16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720104
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>