[nelson][audio][power] Use direct FIDL calls instead of simple codec
For setting AGL in the codec, rather than relying on the simple-codec
library, call the codec FIDL APIs directly. This simplifies the need
for the codec (TAS58xx) driver to support multiple clients (the
brownout driver and the audio engine driver) since each client
using simple-codec will start a hanging get upon initialization which
will cause the second one to get stuck if the server does not support
multiple clients (instances).

simple-codec is being used at the moment for drivers that are not that
simple anymore like the TAS58xx. TAS58xx takes over part of the
simple-codec functionality so it can provide signal processing
capabilities, in this extension it currently does not support multiple
clients like simple-codec does. Rather than adding more complexity in
simple-codec or TAS58xx, we'll scale back the simple-codec functionality
by removing the need for it to handle AGL.

Test: fx test -ov nelson-brownout-protection-test
Change-Id: I31749aaa42a17dbd0cc149798fddf5a023a3ea00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/716440
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>