[thermd] Ensure that logs are always printed on termination
This is workaround for fxb/97657 where logs aren't printed if the
component exits before archivist starts. The workaround is to wait
around whenever we hit an error, which in the case of thermd is any
termination path.

Note that this makes marking thermd as a critical component less useful,
since it will only terminate on a crash, and we may want to do that in
the future. For now, it's more important to log *why* failures occur
than to reboot the device on failure.

Bug: 97657
Bug: 108619
Bug: 108624

Change-Id: Ifb2014dd6b08284bb6bfd6854231ee6f70555343
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719803
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>