[netstack3] Track Address State in Bindings
Change [1/2] in supporting the `WatchAddressAssignmentState` FIDL API.
All relevant integration tests are enabled in the follow up CL.

Propagate updates in `AddressAssignmentState` to the
`AddressStateProvider` Worker when handling `IpDeviceEvents` published
by Core. Updates are sent via an `Unbounded` mpsc channel, which exposes
a synchronized non-blocking send API (required since Core is no-async).

As a result of this change, the initialization logic of the
`AddressStateProvider` worker changed slightly. AddressInfo is added
to `ctx` before the address is added to Core, which ensures the sender
is inplace when Core generates the "Address Added" `IpDeviceEvent`. The
`AddressStateProvider` worker now adds the address to Core as part of
startup.

Bug: 105011

Change-Id: Ica1aa0a4aa5372cc648d9f7fb79160784c31e7fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/715682
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>