[dash-launcher] Add support for custom namespace layouts
This CL adds a new flag to `ffx component explore` called
`-l` or `--layout` that allows changing the namespace
layout given to the dash shell.

The current options are `nested` and `namespace`.

By default, the namespace layout is nested. This means that
instance directories are nested in subdirs at the root
(e.g - component namespace is under /ns, etc.)

When set to `namespace`, the instance namespace is the root.
This most closely resembles what a component instance would
see if it queried its own namespace. This layout is more
compatible with running tools.

To achieve this, we added a new parameter to the Launcher
protocol for selecting the namespace layout.

Bug: 106028
Test: fx test dash_launcher
Change-Id: I3f16cfa262241eeb9ce592d1232d5a6fcd7f3593
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719047
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
API-Review: Gary Bressler <geb@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>