[net-types] Build Cargo features in GN
Create a GN target for the no-std version of net-types (with the "std"
feature disabled). Ensure that both the target with this feature enabled
and the target with it disabled are built and tested in GN. While the
"std" feature is used by all in-tree code, future in-tree code may
disable it, and once we publish on crates.io, external dependents may
disable the it. Thus, it's important to prevent it from bitrotting.

Change-Id: I3f6e9deba96f0b3eb1a45bc5485295892ad89d9d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720041
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>