[policy] Cleanup the structure.
The //src/security/policy directory is our highest traffic directory.
We need to make it as easy as possible to add a new policy without
being confronted by a thousand small files. Almost always the user
is looking for the cfv2 allowlist and very rarely they are looking
for something to do with appmgr. Pkgfs is also a very different
type of allowlist than the other ones and so it deserves its own
distinct directory. This cleanup should make it easier for a new
Fuchsia Engineer who has never worked with policies before just jump
in make a few line edit and be ready to go.

Change-Id: I2e05dfef4677d4054931890123f3d8da63407a75
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720457
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Ben Wright <benwright@google.com>