[benchmarks] Disable input_latency_benchmarks_test due to test ordering problems
Disable input_latency_benchmarks_test because it has been getting
broken by rolls of the test-durations package, which often change test
ordering and sharding.  The cause appears to be that
input_latency_benchmarks_test fails when run after the
screen_is_not_black test.

In this failure cases of this test, we get this error in syslog.txt:

[scenic] ERROR: [src/ui/scenic/lib/input/pointerinjector_registry.cc(105)] InjectorRegistry::Register : Argument |config.context| must be connected to the Scene, and |config.target| must be a descendant of |config.context|

It's possible that is caused by a race condition that is exacerbated
by conditions created when running this test after the
screen_is_not_black test.  Fixing this may involve switching the test
to use Flatland.

Bug: 108167
Change-Id: I22517d3bf9ac147763505416cedd848a9e8e690a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721062
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>