[x86][idle] Don't call Scheduler::Reschedule
Don't call Scheduler::Reschedule when dropping the preempt disable
after waiting up in the idle loop.  Calling Reschedule is a bit of a
waste of effort.  It is going to require us to bounce through the
thread lock, and we know it is not actually going to do anything as we
have explicitly disabled local preemption.

The only thing it is actually going to do is set the local CPU
preemption flag.  We can do that ourselves much more quickly with a
simple relaxed atomic or.  There is no need to make things more
complicated than that.

Change-Id: I79dec862e49acd58da23c17ffd1bba66eff8c8a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717179
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: John Grossman <johngro@google.com>