[netstack] Return error on addr/ep IP mismatch
On Linux, it's not possible to connect to an address with a different IP
version than the endpoint's version. Match this behavior for both V4 and
V6 (the previous code only did so for V4). In addition, add test
coverage for the code path in which we try to connect to a mismatched
IP address when the socket already has an error. This coverage is in
preparation for a later commit that refactors the handling of mismatched
IP addr errors, which (absent coverage) can introduce panics in the
Netstack.

Fixed: 108755
Bug: 104640
Change-Id: Ife0e6fbb7ef702c5559f58057e342bd102af8e91
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719646
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>