[virtualization] Bind to balloon via guest protocol
This change moves connecting to the balloon controller into the guest
protocol. The goal is to have the VMM only expose a single client API,
and a single control plane API.

This also simplifies VMM lifecycle. Before this change we had to ensure
that the VMM was running before binding to the BalloonController (since
it doesn't make sense without a device), but now that's implicitly
enforced by using the guest protocol served by the VMM.

Bug: 104989
Change-Id: Idfa2d85ea4386e3b3e28356c8648a266e2cac3a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720504
Reviewed-by: Alexander Radchenko <alerad@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>