[eager_package_config] Remove test config.
If a build includes the unit test, the test eager package config will
also be included in the build, affecting the real omaha-client and
pkg-resolver behavior.

The purpose of this test config is to test invocation of the script from
GN, now that we have a real eager package config for ota test, this
config can be removed.

Change-Id: I6f6cbfd13da4e98d42f6885a496ba1bc5e9062d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721463
Reviewed-by: James Buckland <jbuckland@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Sen Jiang <senj@google.com>