[docgen] Add an end-to-end test.
Adds some headers that are run through the documentation infrastructure
and compared to golden versions of the generated documentation.

Check in golden files representing the current state of the
documentation generator. Note that these generate .md for devsite to
serve which is not always handled by the Gerrit previewer.

Currently this is not run by default. Once the GN patch to improve
compilation database generation is landed we can enhance this.

Add the capability of the build templates to emit a directory of files.
This can be used because the golden file infrastructure requires that we
know the output names of all files anyway.

Improve error messages of clang doc failures.

Change-Id: Ib328a7684ed83b39e0af70b6935ae7e8b5037a78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721179
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>