[recovery] Log to debuglog
Recovery stopped logging to debuglog when it migrated to CFv2, as CFv2
requires components to be explicit about this. As a result, recovery
logs would no longer appear over serial, which broke the boot simulator
test.

This change uses the stdout_to_debuglog library to allow recovery to log
to debuglog (and serial) again.

Bug: b/241963587
Test: (with follow-up change) fx test -v recovery_simulator_boot_test
Change-Id: I25b0d49011af636c950bafcccf3ce773592846ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/719027
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Eric Stone <ecstone@google.com>