[emu] Add --config to ffx emu show
This CL adds the "--config" flag to the `ffx emu show` subcommand. This
flag outputs the FlagData/Launch Config in the same format as that taken
as input by the "--config" flag for `ffx emu start`. The idea here is
that a user can start an emulator, copy the config using this new flag,
and then feed that config back to a new emulator instance in the future
instead of having to re-build the configuration. This is especially
useful when custom flags are added using the "--edit" functionality.

This CL also moves the existing default to the "--raw" flag, and
repurposes the "--all" flag to print all of the formatted data segments
(excluding the raw output). The raw output remains the default.

Change-Id: I730ca46039eda7bf38f49cfd9aa85242420c58b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721722
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>