[netstack/bench] Don't double record loopback
Currently the netstack iperf3 benchmarks mark the benchmark over
loopback as both a "send" and "recv" benchmark. This is somewhat
misleading because all other metrics use "send" to indicate a test where
the device is sending to the host, and "recv" for the reverse, whereas
in the loopback case "send" and "recv" are from the same test. This also
means that the same metric gets recorded twice.

Instead, record results from iperf3 over loopback just once, and label
each benchmarking setup as one of "send" (DUT-to-host), "recv"
(host-to-DUT), or "loopback" (on-device).

Bug: 108749
Test: fx test --e2e host_x64/netstack_iperf_test
Change-Id: I4831d4ddbf13a9431429ce3bc0df1341c54cf1df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720035
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>