[ktrace] Migrate thread and process info records to FXT
Previously the process name records that are emitted on trace start were
migrated to FXT. This moves the rest of the process name records to FXT,
including those that are emitted for new processes, and the thread
records, both for trace start and for new threads.

Also removes several unused event types around process and thread
lifetimes.

Bug: fxbug.dev/98176
Tested: Start a process while trace is running and verify that process
name is present.

Change-Id: Iebf6079cf11f1c0a6cd02ace1bac3e77084d1626
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/714603
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>