[e2e] Migrate to CF v2 in package_manager_test
This is a reland of https://fxrev.dev/702023. Different to the
original change we switch to using `run-test-suite` in the target
device shell instead of `run` which v1 used. Additionally we
trivially convert the example programs to tests so we can use
`run-test-suite`.

Bug: fxb/103405

Test: fx test --e2e package_manager_test
Multiply: package_manager_test: 100

Change-Id: Iab39d4252e96090d05e663db631893a9a9415e66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/721985
Reviewed-by: Victor Chang <vfcc@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>