[kernel][vm] Restrict AddNewPagesLocked
Forbids the CanOverwriteContent::NonZero option from being passed
to AddNewPagesLocked, removing the need for it to populate a list of
released pages. None of the existing users of AddNewPagesLocked ever
passed in a non-null list, and this path was never exercised.

This simplifies future changes where content other than vm_page_t might
be pre-existing, which would not be able to be placed in a linked list.

Bug: 60238
Change-Id: I685a1d418eb263746836d2c156df7df979812919
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/717586
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>