[netstack3] Implement Clone/Clone2 for TCP sockets
This diverges from the UDP implementation in that it runs all
operations for a TCP socket on the same executor. When a new socket is
created via Clone/Clone2, the requests for that socket are merged in
with the requests from the existing socket. This has the advantage of
not requiring external synchronization (since no other threads are
touching the same socket).

Bug: 108728
Change-Id: Ifa7a03cf17caf2230279d65f7081b1f94316d3ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/720889
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>