[ffx][test] Stabilize output step 3: remove experimental format
This is the third in a series of changes that stabilize the structured
output format produced by ffx test. These changes make breaking changes
to the experiment flags supported by ffx test. These changes are made
in a series of steps to soft transition known users of the ffx test run
command that rely on structured output. Any other uses of ffx test
experimental features may be broken.

In step 3, the test.experimental_structured_output config flag is
removed, which removes the abiliity to write the experimental version of
the output format. Only the stable format may be written.

In previous commits:
 * Step 1 decoupled experiments that are not ready for stabilization
   from test.experimental_structured_output
 * Step 2 updated the test.experimental_structured_output flag to
   switch between experimental and stable formats.

Bug: 81195
Testing: manually verify that the stable output is produced with
  ffx test run <url> --output-directory <dir>

Change-Id: Ie606e5a7ffbdd1e34ae20e22f22d59fbe246705c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/701243
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>