[fidl][hlcpp] unknown interactions server.
Update HLCPP server to be able to handle unknown interactions. The API
used on the server is shared with the client, which means this API is
slightly suboptimal for the server.

For flexible methods, the response callback accepted on the server takes
the Result union, regardless of whether the method uses error syntax. If
the server tries to respond with the transport_err variant, the response
callback panics.

For handling unknown method calls, a handle_unknown_method method is
provided. To prevent this from being called on the Proxy types, this
method is declared as protected, and the implementation provided on the
proxy always panics.

Bug: 88366
Change-Id: Ic75aa7001047995c4b4147064f42b136b0d4ae18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707223
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>