[blackout] add response to test method
Add an expected response to the test fidl method on the controller.
This will allow the harness to confirm that the test step was actually
called. This also adds a duration to the test step - the test waits that
duration to respond. The integration tests should now confirm that the
test step was responded to because of this, and it lets us return errors
up to the point that the duration expires.

As part of the duration implementation, the test steps had to be
refactored to be better async citizens - they were using the std::fs
functions which blocked the executor, so this switches them to use the
async variants provided by fuchsia_fs.

Fixed: 108753
Test: fx test blackout-target-test
Test: ffx storage blackout integration
Change-Id: I85181e61abe793c5a89dde786b98322b4ec0cd8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722125
Reviewed-by: James Sullivan <jfsulliv@google.com>
API-Review: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>