[graphics][compute] Update `component explore` instructions
There was a small change introduced in fxrev.dev/719047 that removes
the need to prefix an executable with `chns`.

Also updated the related `core shard` with new directory names.  This
was necessary to test the components.  It will probably be rebased
away before submission.

Testability: Tested locally on a chromebook.  See updated README.md.
Change-Id: If8f859a4105b65819fb6a67b298eb5a681811d70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722723
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>