[assembly] Add set_source in AssemblyManifest
Sometimes we would like to move an AssemblyManifest, therefore the
source paths of the artifacts need to also change. This CL adds a
set_source function to facilitate this.

This is necessary for the AssemblyManifest to be used in ProductBundles
and for the paths in the AssemblyManifest to be adjusted during the
construction of ProductBundles.

Bug: 107194
Change-Id: I2ff7e85b6c251ceb09b3a9bcf9b87ace9086e646
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718945
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>