[boards][suspend] Fix enable_suspend includes.
Define the enable-suspend-from-board group which includes the kernel
dependencies needed to enable suspend if the enable_suspend flag is set
to true. This group is then always included in the x64-common board.

This gets around the fact that the board is imported before the
enable_suspend flag is set.

Fixed: 109080
Change-Id: Ic67a6bfb449c97002b5a169075be0a62c1fa7c03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723029
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Sarah Pham <smpham@google.com>