[dfv2][driver_manager] Have Node create the Driver
Refactor the code so that the Node is creating the
DriverComponent. This means the Node class encapsulates
more of its own state logic.

Also create a DriverHost interface so that we could support
multiple types of driver hosts in the future.

Change-Id: Ice01e0e74c6e908e2f29a9cdea2c16f4ee2885a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/722018
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>