[netstack3] Drop `Default` trait for `Buffer`
the `Default` trait is only required because of the blanket impl of
`Takeable`, a Buffer is not required to be default constructible and
should only be constructed through `BufferProvider`. Fixed an incorrect
use of `Default::default` for transition from SYN-RCVD to FIN-WAIT-1.
Tests are still allowed to use `Default`. While I am here, I also
dropped `Buffer` as the super trait of `Residual` as it was never used.

Change-Id: Ie24bb7f2042d889750f40badc37f82c8e4514cfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/718417
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>